+++ ext_test_missingfn_82551/source.lua
@@ -1,1 +1,0 @@
-return { popular_manga = function() end }